# docs/mkdocs/api/swagger.yaml
basePath: /
definitions:
...
  api.allocationFileMeta:
    properties:
      approved_by_proposal:
        type: string
      sha256:
        type: string
      status:
        type: string
      type:
        type: string
      uploaded_at:
        type: string
    type: object
  api.allocationListResponse:
    properties:
      allocations:
        items:
          $ref: '#/definitions/api.allocationFileMeta'
        type: array
    type: object
  api.allocationUploadResponse:
    properties:
      sha256:
        type: string
    type: object
  api.auditPubKeyResponse:
    properties:
      public_key:
        type: string
    type: object
  api.auditResponse:
    properties:
      entries:
        items:
          $ref: '#/definitions/ports.AuditEvent'
        type: array
    type: object
...
  api.errorBody:
    properties:
      code:
        type: string
      invariants:
        items:
          $ref: '#/definitions/api.invariantResultJSON'
        type: array
      message:
        type: string
      request_id:
        type: string
    type: object
  api.errorEnvelope:
    properties:
      error:
        $ref: '#/definitions/api.errorBody'
    type: object
...
  api.invariantResultJSON:
    properties:
      invariant:
        type: string
      ok:
        type: boolean
      reason:
        type: string
    type: object
...
  ports.AuditEvent:
    properties:
      event_name:
        description: EventName matches domain.DomainEvent.EventName()
        type: string
      launch_id:
        description: LaunchID scopes the event to a specific launch for per-launch
          filtering.
        type: string
      occurred_at:
        type: string
      payload:
        description: |-
          Payload is the canonical JSON of the domain event, stored as raw JSON
          so audit entries are human-readable without base64 decoding.
        type: object
      prev_hash:
        description: |-
          PrevHash is the SHA-256 hex digest of the previous entry's raw JSON bytes
          (the full marshaled line, signature included). Empty for the first entry.
          Included in the signature of the current entry so it cannot be silently altered.
        type: string
      signature:
        description: |-
          Signature is a base64 Ed25519 signature over the canonical JSON of this entry
          (excluding the signature field itself), signed by the server's audit key.
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Chain launch coordination server. Coordinators and validators authenticate
    via secp256k1 challenge-response (ADR-036)
  title: ChainCoord API
  version: "1.0"
paths:
...
  /audit/pubkey:
    get:
      description: Returns the server's Ed25519 public key for offline audit log signature
        verification.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.auditPubKeyResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/api.errorEnvelope'
      summary: Get audit public key
      tags:
      - audit
...
  /launch/{id}/allocations:
    get:
      parameters:
      - description: Launch UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.allocationListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.errorEnvelope'
      summary: List allocation files
      tags:
      - allocations
  /launch/{id}/allocations/{type}:
    get:
      description: Returns 302 redirect (attestor mode) or streams the raw opaque
        file bytes (host mode).
      parameters:
      - description: Launch UUID
        in: path
        name: id
        required: true
        type: string
      - description: Allocation type
        enum:
        - accounts
        - claims
        - grants
        - authz
        - feegrant
        in: path
        name: type
        required: true
        type: string
      produces:
      - application/octet-stream
      responses:
        "200":
          description: Raw allocation file bytes (host mode)
          schema:
            type: string
        "302":
          description: Redirect to external URL (attestor mode)
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.errorEnvelope'
      summary: Download an allocation file
      tags:
      - allocations
    post:
      consumes:
      - application/json
      - application/octet-stream
      description: |-
        Committee members only.
        Attestor mode (default): register external URL + SHA-256.
        Host mode: upload raw bytes (requires COORD_GENESIS_HOST_MODE=true).
      parameters:
      - description: Launch UUID
        in: path
        name: id
        required: true
        type: string
      - description: Allocation type
        enum:
        - accounts
        - claims
        - grants
        - authz
        - feegrant
        in: path
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.allocationUploadResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.errorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.errorEnvelope'
        "413":
          description: Request Entity Too Large
          schema:
            $ref: '#/definitions/api.errorEnvelope'
      security:
      - BearerAuth: []
      summary: Upload an allocation file
      tags:
      - allocations
  /launch/{id}/audit:
    get:
      description: Returns all audit log entries for a launch. Visibility-gated (same
        rules as GET /launch/{id}).
      parameters:
      - description: Launch UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.auditResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.errorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.errorEnvelope'
      summary: Get audit log
      tags:
      - audit
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Session token obtained from POST /auth/verify. Format: "Bearer <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"